requirements.txt
langchain-mcp-adapters
langgraph
langchain-openai
langchain
python-dotenv
mcp
streamlit
fastapi
uvicorn
requests
pydantic
notion-client
psycopg2
phi
pytest
langchain-core
openai
pandas
plotly
streamlit-plotly-events